requirements.txt
Django==1.8.4
Django-Select2==4.3.1

Pillow==2.9.0
Unidecode==0.04.18
Werkzeug==0.10.4
argparse==1.2.1
cmsplugin-filer==0.10.1
django-appconf==1.0.1
django-classy-tags==0.6.2
django-cms==3.1.2
django-compress==1.0.1
django-compressor==1.5
django-extensions==1.5.5
django-filer==0.9.9
django-image-cropping==1.0.1
django-mptt==0.6.1
django-polymorphic==0.7.1
django-reversion==1.8.5
django-sekizai==0.8.2
django-treebeard==3.0
djangocms-admin-style==0.2.2
djangocms-column==1.5
djangocms-file==0.1
djangocms-flash==0.2.0
djangocms-googlemap==0.3
djangocms-grid==1.2
djangocms-inherit==0.1
djangocms-link==1.6.2
djangocms-oembed==0.5
djangocms-picture==0.1
djangocms-snippet==1.5
djangocms-style==1.5
djangocms-table==1.2
djangocms-teaser==0.1
djangocms-video==0.1
easy-thumbnails==2.2
html5lib==0.999
micawber==0.3.3

six==1.3.0
wsgiref==0.1.2

